normalized state for articles and gifs in the store!!!
