+++ user-auth/react-app/src/App.js
@@ -8,6 +8,7 @@ import ProtectedRoute from "./components/auth/ProtectedRoute";
 import UsersList from "./components/UsersList";
 import User from "./components/User";
 import Articles from "./components/articles/Articles";
+import Content from "./components/content/Content";
 import { authenticate } from "./store/session";
 
 function App() {
@@ -44,7 +45,8 @@ function App() {
         </ProtectedRoute>
         <ProtectedRoute path="/" exact={true} >
           <h1>My Home Page</h1>
-          <Articles />
+          <Content />
+          {/* <Articles /> */}
         </ProtectedRoute>
       </Switch>
     </BrowserRouter>
